+++ src/onboarding/assets/Onboarding4/0_index.ts
@@ -1,0 +1,3 @@
+import IcRightArrow from './rightArrow.svg?react';
+
+export { IcRightArrow };
